[AWARE] Allow configuration of attach/identity event per id
ACTS provides semantics for waiting for specific events - but
not based on the event content. The pop event API with predicate
continues to pop the named events - losing events which do not
match the predicate.

In order to implement the desired semantics - this CL optionally
modifies the name of the event to include the session ID.

Bug: 31940045
Test: existing tests pass, new event names observed
Change-Id: I534d12e8f58995e058c64bcf48458a2ff23251cc