Removed outdated API call so icon is no longer null
The notification was reusing the old notification's icon through

    mNotification.icon

but the old notification's icon was set with setSmallIcon(). This
function call sets the private mSmallIcon Icon object, and not the
icon public int. This means mNotification.icon always returned 0,
even though the icon was set. It has been likely doing this for
years.

Bug: 67412835
Test: act.py -c <config> -tc BleScapApiTest
Test: opening SL4A from the icon does not crash the app.
Change-Id: I581888c4f73da7c278ebe78f817cf2c9887b3b9e