[AWARE] Allow Responders to be configured to accept requests to ANY
Wi-Fi Aware data-path (NDP) setup requires an Initiator and a Responder.
Both Initiator and a Responder are set up with the MAC address of the
peer to which to connect to as well as a security configuration (Open,
PMK, Passphrase).

The original API (27) allowed a Responder to be configured to accept
a connection from anyone (e.g. using a null MAC address). This creates
ambiguous semantics when an NDP already exists. The second Responder
request could refer to the previously setup NDP or to a new one to be
created. We cannot tell the difference up-front.

The framework removes the ability to use ANY, but the facade is now
written to allow this invalid configuration for testing purposes.

Bug: 72175022
Test: integration tests
Change-Id: I146d33e71e0e74bbdf288e915e74517cf30300df