Removed unused EventServer.
This class is code cruft around the EventFacade, and adds no
additional value or features without adding a massive amount of
complexity. The server created here becomes tied to an
RpcReceiverManager (Session), but is hosted as a standalone server,
meaning altough Events will be sent to this server, it effectively
is only the events from a given session, which can only be feasibly
accessed from a JsonRpcServer anyway.

Bug: None
Test: Ran BleScanApiTest
Change-Id: Ib507f9b3bb92057a0d2ada1007b5cf5a3e1946e4