Update sl4a for guava 27.1
Error Prone now detects calling MultiMap.removeAll with a type that
is not compatible with the MultiMap's key, which catches that
mNamedEventObservers.removeAll was being called with an EventObserver,
and not a String.  Add a new removeNamedEventObserver method that
takes the event name to use as the key for
mNamedEventObservers.removeAll.

Bug: 130306229
Test: m checkbuild
Change-Id: I5d83bf18ecfed409f442d7980a4dc97534805711